+++ components/common.js
@@ -18,7 +18,9 @@ const common = {
    */
   async sendMsgTo(loc, msg, type) {
     try {
-      return await Bot[`send${type}Msg`](loc, msg)
+      // 适配trss
+      // return await Bot[`send${type}Msg`](loc, msg)
+      return await Bot[`pick${type}`](loc).sendMsg(msg)
     } catch (err) {
       log.error(`[common.sendMsgTo]发送${type} ${loc}消息错误`, msg, err)
       return false